moded logging to config
