+++ app/config.py
@@ -5,6 +5,10 @@ from flask_marshmallow import Marshmallow
 from flask_sqlalchemy import SQLAlchemy
 from connexion.middleware import MiddlewarePosition
 from starlette.middleware.cors import CORSMiddleware
+import logging
+import sys
+from flask import request
+
 
 basedir = os.path.abspath(os.path.dirname(__file__))
 
@@ -24,6 +28,14 @@ connex_app.add_middleware(
     allow_headers=["*"],
 )
 
+# Configure logging to the console (stdout)
+logging.basicConfig(
+    stream=sys.stdout,  # Output to console
+    level=logging.INFO,  # Logging level
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+)
+logger = logging.getLogger(__name__)
+
 
 # change port to whatever is needed
 PORT = 5555
@@ -41,6 +53,12 @@ app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 app.config['DEBUG'] = True
 app.config['TESTING'] = True
 app.config['UPLOAD_FOLDER'] = UPLOAD_DIR
+
+@connex_app.app.before_request
+def log_request():
+    logger.info(f"Incoming request: {request.method} {request.url}")
+    logger.info(f"Headers: {dict(request.headers)}")
+    logger.info(f"Body: {request.get_data(as_text=True)}")
 
 @app.after_request
 def add_header(response):
